refactor: Add "adapter" suffix to the file names in the adapters directory
Although the directory name where these files are located is "adapters",
however, in order to identify the file function more clearly,
it is still necessary to add "adapter" to the file name.

An example is the source code of Flutter:

~/Flutter/packages/flutter_tools/lib/src/debug_adapters/
flutter_adapter.dart
flutter_test_adapter.dart
