+++ src/ui/flutter_app/lib/services/storage/adapters/src/algorithm_adapter.dart
