+++ src/ui/flutter_app/lib/services/storage/adapters/src/color_adapter.dart
