+++ src/ui/flutter_app/lib/services/storage/adapters/src/painting_style_adapter.dart
